requirements.txt
pyDataverse
python-dotenv
requests
fastapi
uvicorn
